backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
requests==2.31.0
discord.py==2.3.2
Werkzeug==3.0.1

# Excel processing
openpyxl==3.1.2
xlrd==2.0.1
pandas==2.1.4
